+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2021-04-21
 
+### Added
+
+- new `syntax` config key -- supports: `"twig"`
+
 ### Fixed
 
 - file system pathing issues
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CSV file generation
 - JSON file generation
+- `project` config key -- replaced with `syntax`
 
 ## [0.1.0] - 2020-01-18